requirements.txt
Flask
requests
opencv-python
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openpyxl
pandas
